2019 - Day 09

+++ 2019-rust2024/src/day_02.rs
@@ -1,8 +1,8 @@
 use crate::intcode::{IntComputer, RetCode};
 
 #[must_use]
-pub fn run() -> (isize, isize) {
-    let input: Vec<isize> = include_str!("../input/day02.txt")
+pub fn run() -> (i128, i128) {
+    let input: Vec<i128> = include_str!("../input/day02.txt")
         .strip_suffix("\n")
         .unwrap()
         .split(',')
